Extract `LowPolynomial` to a separate file
